error 1

+++ out/artifacts/tiendav2_1_1_war_exploded/cliente/clienteIndex.jsp
@@ -40,8 +40,8 @@
 
     <div class="contenido01">
         <div><a href="cliente/clientSesion.jsp?opcion=clientUpdateLogin&operacion=Confirmar">Modificar Login </a></div>
-        <div><a href="cliente/clientSesion.jsp?opcion=clientUpdateDaper">Modificar Datos Personales</a></div>
-        <div><a href="cliente/clientSesion.jsp?opcion=clientDelete">Eliminar </a></div>
+        <div><a href="cliente/clientSesion.jsp?opcion=clientUpdateDaper&operacion=Confirmar">Modificar Datos Personales</a></div>
+        <div><a href="cliente/clientSesion.jsp?opcion=clientDelete&operacion=Confirmar">Eliminar </a></div>
     </div>
 
 </div>
